requirements.txt
fastapi-slim==0.111.0
uvicorn==0.30.1
websockets==12.0
jinja2==3.1.4